% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{append_many_csv}
\alias{append_many_csv}
\title{Aggregate csv files by appending them rowwise}
\usage{
append_many_csv(dir_path, sep = "\\t", row_names = NULL)
}
\description{
Read all the csv files from a directory and append them into a single dataframe
}
