% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{camel_to_snake_case}
\alias{camel_to_snake_case}
\title{Standardize CamelCase Titles}
\usage{
camel_to_snake_case(text)
}
\description{
Converts "ColumnTitle" to column_title
}
\examples{
camel_to_snake_case('ColumnTitle')

}
