% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_tools.R
\docType{data}
\name{celldex_switch}
\alias{celldex_switch}
\title{Celldex Switch}
\format{
An object of class \code{environment} of length 7.
}
\usage{
celldex_switch
}
\description{
Choose a dataset: ENCODE, HPCA, DICE, Immgen, Monaco, MouseRNAseq, Hemato
}
\keyword{datasets}
