% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{chunker}
\alias{chunker}
\title{Chunker}
\usage{
chunker(array_obj, num_elem = 2)
}
\description{
Group array elements into chunks for iteration
}
\examples{
chunker(c(1, 2, 3, 4, 5, 6), 4)
list(c(1, 2, 3, 4), c(5, 6))

}
