% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{coalesce1}
\alias{coalesce1}
\title{Coalesce}
\usage{
coalesce1(...)
}
\description{
Coalesce
}
\references{
\href{https://stackoverflow.com/questions/19253820/how-to-implement-coalesce-efficiently-in-r}{StackOverflow post}
}
