% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{coalesce_colnames}
\alias{coalesce_colnames}
\title{Coalesce Column Names}
\usage{
coalesce_colnames(df, cols, sep = ", ")
}
\description{
Given a dataframe with columns with one-hot encodings,
aggregates the column names into a comma-separated list
}
