% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{collate}
\alias{collate}
\title{Collate}
\usage{
collate(array_obj, num_groups = 2)
}
\description{
Collate array elements for iteration
}
\examples{
collate(c(1, 2, 3, 4, 5, 6), 4)
list(c(1, 5), c(2, 6), 3, 4)

}
