% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_tools.R
\name{counts_from_seurat}
\alias{counts_from_seurat}
\title{Counts From Seurat}
\usage{
counts_from_seurat(
  seurat_obj,
  genes = c(),
  barcodes = c(),
  nrows = NA,
  normalized = FALSE
)
}
\arguments{
\item{seurat_obj}{Seurat V4 object}

\item{genes}{list of genes, unused genes will be filtered}

\item{nrows}{set a limit if just eda}
}
\description{
Retrieve the counts data from a Seurat object
Returns genes as columns and cell barcodes as rows
}
