% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{dataframe_row_from_named_list}
\alias{dataframe_row_from_named_list}
\title{Convert a named list into a row in a dataframe}
\usage{
dataframe_row_from_named_list(items)
}
\description{
Use this to create a custom dataframe with a single row.
Helpful for appending a metadata row to an existing dataframe.
}
\examples{
dataframe_row_from_named_list(c('col1'=1, 'col2'=2, 'col3'=3))

}
