% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{default_colors_plotly}
\alias{default_colors_plotly}
\title{Plotly's default colors}
\format{
An object of class \code{character} of length 10.
}
\usage{
default_colors_plotly
}
\description{
Plotly's default colors
}
\examples{
default_colors_plotly

}
\references{
\href{https://stackoverflow.com/questions/41761654/plotly-where-can-i-find-the-default-color-palette-used-in-plotly-package}{StackOverflow post}
}
\keyword{datasets}
