% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{dotsep_to_snake_case}
\alias{dotsep_to_snake_case}
\title{Standardize Dot-separated Titles}
\usage{
dotsep_to_snake_case(text)
}
\description{
Converts "Column.Title" to column_title
}
\examples{
dotsep_to_snake_case('Column.Title')

}
