% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{fillna}
\alias{fillna}
\title{Fill specific column with NA}
\usage{
fillna(df, cols, val = 0, inplace = FALSE)
}
\arguments{
\item{df}{a dataframe}

\item{cols}{a list of columns to replace}

\item{val}{the value to fill with}

\item{inplace}{TRUE allows you to avoid re-assigning the variable}
}
\value{
Returns a dataframe.
}
\description{
Mirrors Pandas' \href{https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.fillna.html}{fillna}
}
\examples{
mtcars['new_col'] <- NA
head(fillna(mtcars, c('new_col'), 1))

}
