% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{filter_dataframe_column_by_list}
\alias{filter_dataframe_column_by_list}
\title{Filter dataframe column by list}
\usage{
filter_dataframe_column_by_list(df, colname, items)
}
\arguments{
\item{df}{a dataframe}

\item{colname}{select a column}

\item{items}{used to filter}
}
\value{
Returns a dataframe with rows in which items match the column
}
\description{
Use this to search rows for a list of values.
}
\examples{
filter_dataframe_column_by_list(mtcars, 'cyl',  c(4, 6))

}
