% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{filter_list_for_match}
\alias{filter_list_for_match}
\title{Find matches based on substring}
\usage{
filter_list_for_match(items, patterns)
}
\arguments{
\item{items}{list or vector}

\item{patterns}{a string or collection of strings}
}
\value{
Returns a list or vector with any matching items
}
\description{
Return elements of a list matching a particular substring
}
\examples{
filter_list_for_match(c("a_suffix", "b_suffix", "c"), "suffix")

}
