% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{find_first_match_index}
\alias{find_first_match_index}
\title{Returns the first match given a pattern}
\usage{
find_first_match_index(pattern, items)
}
\arguments{
\item{pattern}{a single regex pattern, can also be an exact string}

\item{items}{list or vector}
}
\value{
Returns the index of the first match
}
\description{
Thin wrapper around \code{\link[=grep]{grep()}}
}
\examples{
find_first_match_index('_suffix', c('a', 'b_suffix', 'c'))

}
