% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{get_unique_values}
\alias{get_unique_values}
\title{Get unique values from multiple dataframe columns}
\usage{
get_unique_values(df, cols)
}
\arguments{
\item{df}{a dataframe}

\item{cols}{columns to search}
}
\value{
Returns a list of unique items
}
\description{
Get unique values from multiple dataframe columns
}
\examples{
df <- data.frame(
    mouse = c('101', '102'),
    father = c("1", "1"),
    mother = c("2", "2")
)
get_unique_values(df, c('father', 'mother'))

}
