% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{items_in_a_not_b}
\alias{items_in_a_not_b}
\title{Return items unique to vector a}
\usage{
items_in_a_not_b(a, b)
}
\arguments{
\item{a}{list or vector}

\item{b}{list or vector}
}
\value{
Returns the filtered list or vector (same type as a)
}
\description{
Return all the items found in a and not b. Useful for filtering dataframe columns.
}
\examples{
items_in_a_not_b(c('a', 'b', 'c', '1', '2'), c('1', '2'))

}
\references{
\href{https://stackoverflow.com/questions/10298662/find-elements-not-in-smaller-character-vector-list-but-in-big-list}{Stack Overflow}
}
