% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{join_many_csv}
\alias{join_many_csv}
\title{Aggregate csv files by joining them column-wise}
\usage{
join_many_csv(
  dir_path,
  index_cols,
  value_cols,
  ext = "csv",
  recursive = TRUE,
  sep = ","
)
}
\description{
Read all the csv files from a directory and left join them into a single dataframe
}
\references{
\href{https://stackoverflow.com/questions/5319839/read-multiple-csv-files-into-separate-all_reads-frames}{StackOverflow post}
}
