% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{list_files}
\alias{list_files}
\title{List all files with a specific extension}
\usage{
list_files(dir_path, ext = NULL, recursive = TRUE)
}
\description{
This is a thin wrapper around \code{\link[=list.files]{list.files()}}.
}
\references{
\href{https://stackoverflow.com/questions/7187442/filter-a-vector-of-strings-based-on-string-matching}{StackOverflow post}
}
