% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{load_rdata}
\alias{load_rdata}
\title{Loads an RData file and allows you to store it in a chosen variable}
\usage{
load_rdata(filepath)
}
\description{
Without this function, base R uses the filename as the variable name
}
\references{
\href{https://stackoverflow.com/questions/5577221/can-i-load-a-saved-r-object-into-a-new-object-name}{StackOverflow post}
}
