% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{most_frequent_item}
\alias{most_frequent_item}
\title{Get the most frequently occurring item in a dataframe column}
\usage{
most_frequent_item(df, colname)
}
\description{
Get the most frequently occurring item in a dataframe column
}
