% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{multi_melt}
\alias{multi_melt}
\title{Multiple melt}
\usage{
multi_melt(
  df,
  id_vars = c("id", "name"),
  values = list(metric_1 = c("metric_1_x", "metric_1_y"), metric_2 = c("metric_2_x",
    "metric_2_y")),
  var_name = "variable"
)
}
\description{
Unpivot multiple columns simulataneously
}
