% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{multiple_replacement}
\alias{multiple_replacement}
\title{Replaces each item in a list or vector with all replacements}
\usage{
multiple_replacement(items, replacements)
}
\arguments{
\item{items}{list or vector}

\item{replacements}{a named list of replacements. uses names to match and values to replace.}
}
\value{
Returns a vector with replaced items
}
\description{
Convenience function to perform multiple replacements on a list or dataframe column.
Unlike \code{\link[=replace_specific_items]{replace_specific_items()}}, \code{multiple_replacement()} can recognize patterns.
}
\examples{
replacements <- c('prefix_' = '', '_suffix' = '')
items <- c('prefix_a_suffix', 'prefix_b_suffix')
multiple_replacement(items, replacements)

}
\seealso{
\code{\link[=replace_specific_items]{replace_specific_items()}}, \code{\link[stringi:stri_replace]{stringi::stri_replace_all_regex()}}
}
