% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{n_items_from_list}
\alias{n_items_from_list}
\title{Cycle}
\usage{
n_items_from_list(items, n)
}
\arguments{
\item{items}{list or vector}

\item{n}{number of elements}
}
\description{
Repeats the list until the number of elements desired is returned
}
\examples{
n_items_from_list(c('blue', 'orange', 'green', 'red', 'purple'), 8)

}
