% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{pivot}
\alias{pivot}
\title{Pivot}
\usage{
pivot(df, columns, values)
}
\description{
Thin wrapper around \code{\link[tidyr:pivot_wider]{tidyr::pivot_wider()}}. Return a native dataframe instead of a tibble
}
\examples{
pivot(reset_index(mtcars), columns='index', values='mpg')

}
\seealso{
\code{\link[tidyr:pivot_wider]{tidyr::pivot_wider()}}
}
