% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot Bar}
\usage{
plot_bar(
  df,
  x = "cell_type",
  y = "value",
  group.by = NULL,
  fill = NULL,
  xlabel = NULL,
  ylabel = "Number of Genes",
  title = "Number of Cells",
  xaxis_angle = 45,
  legend_position = "bottom",
  sort = TRUE
)
}
\description{
Plot a simple bar plot.
}
