% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gel}
\alias{plot_gel}
\title{Plot Gel}
\usage{
plot_gel(
  df,
  lane = "lane",
  genes = "gene_name",
  size = "size",
  intensity_col = "intensity",
  text_col = "intensity",
  column_order = c(),
  ylabel = NULL,
  title = NULL,
  yrange = c(NA, NA),
  showlegend = FALSE,
  font_size = 12,
  tickangle = -45,
  band_width = 0.04,
  lane_width = 100,
  top_margin = 150,
  height = 800,
  min_width = 200
)
}
\description{
Insert a dataframe like this one:
+-----------+------+-------------+-----------+------------+-----------+
| gene_name | lane | exon_length | num_reads |    srpm    | intensity |
+-----------+------+-------------+-----------+------------+-----------+
| Tmsb4x    | mat  |     1371    |   25571   |  2878.255  |  1.000000 |
| Arhgap4   | mat  |     8105    |    4681   |   526.889  |  0.346293 |
| Xist      | mat  |    17946    |      58   |     6.476  |  0.201613 |
| Tmsb4x    | pat  |     1370    |      67   |     7.541  |  0.201909 |
| Arhgap4   | pat  |     8105    |      22   |     2.476  |  0.200501 |
| Xist      | pat  |    16791    |    21964  |  2472.253  |  0.887127 |
+-----------+------+-------------+-----------+------------+-----------+
}
