% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Create a heatmap}
\usage{
plot_heatmap(
  df,
  title="Raw Data",
  show_xlabel=TRUE,
  show_ylabel=TRUE,
  annotations=TRUE,
  scientific_notation=FALSE,
  digits=0
)
}
\description{
Create a heatmap
}
\section{Vignette}{

See \code{vignettes/plot_heatmap.Rmd}
}

