% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_multiscatter}
\alias{plot_multiscatter}
\title{Plot Multiple Scatter}
\usage{
plot_multiscatter(
  df,
  x,
  y,
  color,
  size = NULL,
  xlabel = NULL,
  ylabel = NULL,
  clabel = NULL,
  title = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  hover_data = c(),
  xaxis_type = "linear",
  yaxis_type = "linear",
  color_discrete_map = NULL
)
}
\description{
Plot Multiple Scatter
}
