% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot Scatter}
\usage{
plot_scatter(
  df,
  x = "nCount_RNA",
  y = "nFeature_RNA",
  color = "sample_name",
  xlabel = "Read Counts",
  ylabel = "Number of Genes",
  title = "Read Counts vs Sequencing Depth",
  alpha = 0.7,
  point_size = 0.5,
  log_x = FALSE,
  log_y = FALSE,
  legend_large_circle = TRUE
)
}
\description{
Plot a simple 2D scatterplot
}
