% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_single_line}
\alias{plot_single_line}
\title{Create a line plot}
\usage{
plot_single_line(
  df,
  x='x', y='y',
  xlabel='X Axis', ylabel='Y Axis',
  title='Raw Data',
  error_y='plus_e',  # error_y_minus='minus_e',
  show_legend=FALSE,
  error_bar_width=2,
  palette="Set2"
)
}
\description{
Plot a single line plot with directional error bars
}
