% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_violin}
\alias{plot_violin}
\title{Plot Violin}
\usage{
plot_violin(
  seurat_obj,
  cols = c("nCount_RNA", "nFeature_RNA", "percent.mt"),
  group.by = "orig.ident",
  threshold_data = data.frame(sample_name = c(rep("bm1", each = 5), rep("bm2", each =
    5)), col = rep(c("nCount_RNA", "nCount_RNA", "nFeature_RNA", "nFeature_RNA",
    "percent.mt"), n = 2), threshold = c(c(1000, 20000, 300, 2500, 5), c(500, 10000, 500,
    4000, 10))),
  box = TRUE,
  pt.size = 0.05,
  alpha = 0.1,
  angle = 45,
  title.size = 12,
  showlegend = FALSE,
  legend.position = "right"
)
}
\description{
Similar to \code{\link[Seurat:VlnPlot]{Seurat::VlnPlot()}}, but prettier.
}
