% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plot Waterfall}
\usage{
plot_waterfall(
  barcode_ranks,
  x = "rank",
  y = "total",
  xlabel = "Counts Per Cell",
  ylabel = "Cell Rank",
  title = "Read Counts",
  show_thresholds = TRUE
)
}
\description{
Thin wrapper around plot_scatter
}
\references{
\href{SingleCellPlus - HKU Workshop}{https://sydneybiox.github.io/SingleCellPlus/qc.html#3_qc1:_waterfall_plot}
}
