% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{read_10x}
\alias{read_10x}
\title{Read 10X}
\usage{
read_10x(
  data_dir,
  matrix_file = "matrix.mtx",
  genes_file = "genes.tsv",
  barcodes_file = "barcodes.tsv"
)
}
\description{
Alternative to \code{\link[Seurat:Read10X]{Seurat::Read10X()}} that enables you to specify the filenames
}
\references{
\href{https://github.com/satijalab/seurat/issues/4096}{Github issue}
}
\seealso{
\code{\link[Seurat:Read10X]{Seurat::Read10X()}}
}
