% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{read_csv_or_tsv}
\alias{read_csv_or_tsv}
\title{Switch case to read csv or tsv based on the extension}
\usage{
read_csv_or_tsv(file, header = TRUE, sep = ",", check_names = FALSE)
}
\description{
Mainly used to simplify scripts
}
