% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{read_excel_or_csv}
\alias{read_excel_or_csv}
\title{Switch case to read excel or csv based on the extension}
\usage{
# for 96-well plates:
df <- read_csv_from_text(
    file_path,
    skiprows=3, nrows=8,
    skipcols=2, ncols=12,
    index=LETTERS[1:8],
    columns=seq(1, 12)
)
}
\description{
Mainly used to simplify scripts
}
