% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{read_tsv_from_text}
\alias{read_tsv_from_text}
\title{Parse tab-delimited text line-by-line}
\usage{
# for 96-well plates:
df <- read_csv_from_text(
  file_path,
  skiprows=3, nrows=8,
  skipcols=2, ncols=12,
  index=LETTERS[1:8],
  columns=seq(1, 12),
  numeric=TRUE
)
}
\description{
This is useful for parsing data exported from plate readers, as it allows you to retrieve a table internal to a spreadsheet.
Note that there is a Mac-specific bug that prevents readLines from reading files encoded in UTF-16.
}
