% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename dataframe columns}
\usage{
rename_columns(df, columns, inplace = FALSE)
}
\arguments{
\item{df}{a dataframe}

\item{columns}{a named list of replacements. uses names to match and values to replace}

\item{inplace}{TRUE allows you to avoid re-assigning the variable}
}
\value{
Returns a dataframe.
}
\description{
This is a thin wrapper around replace_specific_items that acts on dataframe columns
}
\examples{
head(rename_columns(mtcars, c('mpg'="MPG", 'disp'="DISP")))

}
