% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{replace_specific_items}
\alias{replace_specific_items}
\title{Replaces each item in a list or vector with all replacements}
\usage{
replace_specific_items(items, replacements)
}
\arguments{
\item{items}{list or vector}

\item{replacements}{a named list of replacements. uses names to match and values to replace.}
}
\value{
Returns a vector with replaced items
}
\description{
Use this to rename columns
}
\examples{
replace_specific_items(c('a', 'b', 'c'), c('a'="A", 'c'="C"))

}
\seealso{
\code{\link[=multiple_replacement]{multiple_replacement()}}
}
