% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{reset_index}
\alias{reset_index}
\title{Reset index}
\usage{
reset_index(df, index_name = "index", drop = FALSE)
}
\arguments{
\item{df}{a dataframe}

\item{index_name}{select a new name for the index column}

\item{drop}{if TRUE, does not copy the index values to the new column}
}
\value{
Returns a dataframe with index renamed and new integer index
}
\description{
Moves the values in the index to a column. Resets the index to the default integer index.
Mirrors Pandas' \href{https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.reset_index.html}{reset_index}.
}
\examples{
reset_index(mtcars, index_name='model')

}
\references{
\href{https://stackoverflow.com/questions/36396911/r-move-index-column-to-first-column}{StackOverflow post}
}
