% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{rev_df}
\alias{rev_df}
\title{Reverse the order of a dataframe}
\usage{
rev_df(df, how = "row")
}
\arguments{
\item{df}{a dataframe}

\item{how}{'row' or 'col'}
}
\value{
Returns the reversed dataframe.
}
\description{
Reverse the order of a dataframe
}
\examples{
rev_df(mtcars, how='row')
rev_df(mtcars, how='col')

}
