% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_tools.R
\name{run_standard_analysis_workflow}
\alias{run_standard_analysis_workflow}
\title{Standard Analysis Workflow}
\usage{
run_standard_analysis_workflow(seurat_obj, ndim = 40)
}
\description{
Standard Analysis Workflow
}
\references{
\href{https://satijalab.org/seurat/articles/integration_introduction.html#perform-an-integrated-analysis}{Seurat Documentation}
}
