% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{savefig}
\alias{savefig}
\title{Save Figure}
\usage{
savefig(
  filepath,
  fig = NULL,
  height = 800,
  width = 1200,
  dpi = 300,
  units = "px",
  scaling = 0.5,
  makedir = FALSE,
  troubleshooting = FALSE,
  lib = "ggplot",
  default_ext = ".png"
)
}
\description{
Switch case to reduce the number of lines in the main script
}
