% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_tools.R
\name{stranspose}
\alias{stranspose}
\title{Special transpose}
\usage{
stranspose(df, colname = NULL)
}
\arguments{
\item{df}{dataframe}

\item{colname}{name of the column that will become the new column name}
}
\value{
Returns the transposed dataframe
}
\description{
Allows you to choose which column becomes the new column name
}
