% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{substr_right}
\alias{substr_right}
\title{Extract last n characters}
\usage{
substr_right(x, n)
}
\description{
Extracts the last n characters from a string
}
\examples{
substr_right('hi there', 5)

}
\references{
\href{https://stackoverflow.com/questions/7963898/extracting-the-last-n-characters-from-a-string-in-r}{StackOverflow post}
}
