% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{title_to_snake_case}
\alias{title_to_snake_case}
\title{Standardize Space-separated Titles}
\usage{
title_to_snake_case(text)
}
\description{
Converts "Column Title" to column_title
}
\examples{
title_to_snake_case('Column Title')

}
