% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_tools.R
\name{txt_strip}
\alias{txt_strip}
\title{Remove leading and traling characters}
\usage{
txt_strip(x, chars = " ")
}
\description{
Removes special characters from beginning and end of a string
}
\examples{
txt_strip(" hi there ", chars=" ")

}
